refactor: rename to computeCursorArgFilter, align naming with backend, inline resolveOrderByFields
